requirements.txt
requests
python-dotenv
rich
gpt4all
openai
transformers
selenium
webdriver-manager